fix: perbaikan pada loop agar tampil

+++ tasks/index.php
@@ -8,6 +8,7 @@ if (!isset($_SESSION['user_id'])){
 
 $user_id = $_SESSION['user_id'];
 $result = $conn->query("SELECT * FROM tasks WHERE user_id = $user_id");
+$result->fetch_assoc();
 
 ?>
 
@@ -16,11 +17,12 @@ $result = $conn->query("SELECT * FROM tasks WHERE user_id = $user_id");
 
 <h2>Task Saya</h2>
 
-<?php while($row = $result->fetch_assoc()); ?>
+<?php foreach($result as $row): ?>
     <div>
         <strong><?= $row['title'] ?></strong>
-        <?= $row['active'] ? '[AKTIF]' : '[NONAKTIF]' ?>
+        <?= $row['is_active'] ? '[AKTIF]' : '[NONAKTIF]' ?>
         <P><?= $row['description'] ?></P>
-        <a href="edit.php?id=<? $row['id'] ?>">edit</a>
-        <a href="delete.php?id=<? $row['id'] ?>">hapus</a>
+        <a href="edit.php?id=<?= $row['id'] ?>">edit</a>
+        <a href="delete.php?id=<?= $row['id'] ?>" onclick="return confirm('yakin nih mo diapus')">hapus</a>
     </div>
+<?php endforeach; ?>
